Refactoring optimizer by adding learning rate schedule
