+++ python/ml4ir/applications/ranking/tests/data/config/model_config.yaml
@@ -1,0 +1,29 @@
+architecture_key: dnn
+layers:
+  - type: dense
+    name: first_dense
+    units: 256
+    activation: relu
+  - type: dropout
+    name: first_dropout
+    rate: 0.0
+  - type: dense
+    name: second_dense
+    units: 64
+    activation: relu
+  - type: dropout
+    name: second_dropout
+    rate: 0.0
+  - type: dense
+    name: final_dense
+    units: 1
+    activation: null
+optimizer:
+  key: sgd
+  gradient_clip_value: 5.0
+lr_schedule:
+  key: cyclic
+  type: triangular2
+  initial_learning_rate: 0.07
+  maximal_learning_rate: 0.7
+  step_size: 7
