+++ python/ml4ir/base/model/optimizers/cyclic_learning_rate.py
@@ -1,23 +1,9 @@
-# Copyright 2019 The TensorFlow Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Cyclical Learning Rate Schedule policies for TensorFlow."""
+"""Cyclical Learning Rate Schedule policies for TensorFlow.
+Adopted from:
+https://www.tensorflow.org/addons/api_docs/python/tfa/optimizers/CyclicalLearningRate
+"""
 
 import tensorflow as tf
-#from tensorflow_addons.utils.types import FloatTensorLike
-
-#from typeguard import typechecked
 from typing import Union, Callable
 
 
@@ -51,7 +37,9 @@ class CyclicalLearningRate(tf.keras.optimizers.schedules.LearningRateSchedule):
         ```
         You can pass this schedule directly into a
         `tf.keras.optimizers.Optimizer` as the learning rate.
-        Args:
+
+        Parameters
+        ----------
             initial_learning_rate: A scalar `float32` or `float64` `Tensor` or
                 a Python number.  The initial learning rate.
             maximal_learning_rate: A scalar `float32` or `float64` `Tensor` or
@@ -62,7 +50,10 @@ class CyclicalLearningRate(tf.keras.optimizers.schedules.LearningRateSchedule):
             scale_mode: ['cycle', 'iterations']. Mode to apply during cyclic
                 schedule
             name: (Optional) Name for the operation.
-        Returns:
+
+
+        Returns
+        -------
             Updated learning rate value.
         """
         super().__init__()
@@ -86,9 +77,11 @@ class CyclicalLearningRate(tf.keras.optimizers.schedules.LearningRateSchedule):
 
             mode_step = cycle if self.scale_mode == "cycle" else step
 
-            return initial_learning_rate + (
+            lr = initial_learning_rate + (
                 maximal_learning_rate - initial_learning_rate
             ) * tf.maximum(tf.cast(0, dtype), (1 - x)) * self.scale_fn(mode_step)
+
+            return lr
 
     def get_config(self):
         return {
