+++ python/ml4ir/base/pipeline.py
@@ -112,7 +112,6 @@ class RelevancePipeline(object):
 
         # Setup other arguments
         self.loss_key: str = self.args.loss_key
-        self.optimizer_key: str = self.args.optimizer_key
         if self.args.metrics_keys[0] == "[":
             self.metrics_keys: List[str] = ast.literal_eval(self.args.metrics_keys)
         else:
